New stuff

+++ Food-Analyzer_Project/src/json/extractor/food/nutrient/FoodNutrients.java
@@ -2,5 +2,5 @@ package json.extractor.food.nutrient;
 
 import java.util.List;
 
-public record FoodNutrients(Nutrient nutrient, double amount) {
+public record FoodNutrients(Nutrient nutrient, float amount) {
 }
